pyproject.toml
[build-system]
requires = ["hatchling", "hatch-fancy-pypi-readme"]
build-backend = "hatchling.build"

[project]
name = "libertem-blobfinder"
dynamic = ["version", "readme"]
description = "LiberTEM correlation and refinement library"
requires-python = ">=3.9"
license-files = { paths = ["LICENSE"] }
dependencies = [
    "numpy",
    "numba>=0.49",
    "sparseconverter",
    "scikit-image",
]
keywords = ["electron microscopy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Visualization",
]
authors = [
    { name = "the LiberTEM team", email = "libertem-dev@googlegroups.com" },
]

[project.optional-dependencies]
udf = ["libertem>=0.12.0,<1", "matplotlib"]
common = ["libertem>=0.12.0,<1"]
hdbscan = ["hdbscan"]

[project.urls]
Repository = "https://github.com/LiberTEM/LiberTEM-blobfinder"
Documentation = "https://libertem.github.io/LiberTEM-blobfinder"

[tool.hatch.metadata.hooks.fancy-pypi-readme]
"content-type" = "text/x-rst"

[[tool.hatch.metadata.hooks.fancy-pypi-readme.fragments]]
path = "README.rst"

[[tool.hatch.metadata.hooks.fancy-pypi-readme.substitutions]]
pattern = ":(cite|doc):`[^`]+` ?"
replacement = ""

[tool.hatch.version]
path = "src/libertem_blobfinder/__version__.py"

[tool.hatch.build.hooks.custom]
# this enables hatch_build.py
